% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_algorithm.R
\name{add_diagnoses}
\alias{add_diagnoses}
\title{Add diagnoses to the input dataset}
\usage{
add_diagnoses(banff_dataset, version = NULL)
}
\arguments{
\item{banff_dataset}{A tibble object.}

\item{version}{A character string referring the version of Banff classification.
The most recent classification is the default.
Options are "2022" (default), "2017".}
}
\value{
A tibble object that contains additional variables with the diagnoses results.
}
\description{
This function takes a dataset and returns a diagnosis for each
observation. For the function to run, the dataset must not contain any errors
that \code{\link[=banff_launcher]{banff_launcher()}} would have detected. Please prefer using
\code{\link[=banff_launcher]{banff_launcher()}} to run additional tests.
}
\examples{
{

banff_dataset <- get_banff_template()
add_diagnoses(banff_dataset)

}

}
