% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{banffIT_website}
\alias{banffIT_website}
\title{Call the online documentation}
\usage{
banffIT_website()
}
\value{
Nothing to be returned. The function opens a web page.
}
\description{
This function sends the user to the online documentation for the package,
which includes a description of the latest version of the package, vignettes,
user guides, and a reference list of functions and help pages.
}
\examples{
{

banffIT_website()

}

}
