% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_algorithm.R
\name{calculate_adequacy}
\alias{calculate_adequacy}
\title{Calculate adequacy of each biopsy from glomeruli and arteries variables}
\usage{
calculate_adequacy(banff_dataset)
}
\arguments{
\item{banff_dataset}{A tibble object.}
}
\value{
A tibble object with two variables: the calculated adequacy
(adequacy_calculated) and the adequacy specified in input (adequacy_input).
}
\description{
This function calculates adequacy of each biopsy (i.e., each observation)
based on glomeruli and arteries variables (if provided).
}
\examples{
{

banff_dataset <- get_banff_template()
calculate_adequacy(banff_dataset)

}

}
