% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{get_banff_template}
\alias{get_banff_template}
\title{Get a dataset template}
\usage{
get_banff_template(version = NULL)
}
\arguments{
\item{version}{A character string referring the version of Banff classification.
The most recent classification is the default.
Options are "2022" (default), "2017".}
}
\value{
A tibble representing the empty dataset.
}
\description{
This function gets the empty dataset with variables that are mandatory in the
process.
}
\examples{
{

  get_banff_template()

}

}
