% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{get_banff_version}
\alias{get_banff_version}
\title{Extract version of the Banff classification}
\usage{
get_banff_version(version = NULL)
}
\arguments{
\item{version}{A character string referring the version of Banff classification.
The most recent classification is the default.
Options are "2022" (default), "2017".}
}
\value{
A character string which is the version of Banff classification.
}
\description{
This internal function extracts for user input the version of the Banff
classification.
}
\examples{
{

get_banff_version()

}

}
