% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_algorithm.R
\name{add_diagnoses}
\alias{add_diagnoses}
\title{Add diagnoses to a Banff dataset}
\usage{
add_diagnoses(banff_dataset)
}
\arguments{
\item{banff_dataset}{A tibble object.}
}
\value{
A tibble object that contains additional columns with the diagnoses results.
}
\description{
This function takes a dataset and returns a diagnosis for each
observation. For the function to run, the dataset must not contain any errors
that \code{\link[=banff_launcher]{banff_launcher()}} would have detected. Please prefer using
\code{\link[=banff_launcher]{banff_launcher()}} to run additional tests.
}
\examples{
{

library(fabR)
input_file <- system.file("extdata", "example.xlsx", package = "banffIT")
banff_dataset <- read_excel_allsheets(input_file)[1,]
add_diagnoses(banff_dataset)

}

}
