% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_algorithm.R
\name{add_diagnosis}
\alias{add_diagnosis}
\title{Add diagnosis to a banff dataset}
\usage{
add_diagnosis(banff_dataset, .test_TA = FALSE)
}
\arguments{
\item{banff_dataset}{A tibble object.}

\item{.test_TA}{Temoporary parameter to test with 'ta' variable.}
}
\value{
A tibble object which contain additional columns refering to diagnosis.
}
\description{
This function takes in input a dataset which is given for all the observation
a diagnosis. To be able to perform, the dataset must not contain any errors
that \code{\link[=banff_launcher]{banff_launcher()}} would not have detected. Please prefer using
\code{\link[=banff_launcher]{banff_launcher()}} to effectuate additional tests.
}
\examples{
{

library(fabR)
banff_file <- system.file("extdata", "example.xlsx", package = "banffIT")
banff_dataset <- read_excel_allsheets(banff_file)
add_diagnosis(banff_dataset)

}

}
