% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_algorithm.R
\name{banff_dataset_evaluate}
\alias{banff_dataset_evaluate}
\title{Assesses the banff dataset}
\usage{
banff_dataset_evaluate(banff_dataset)
}
\arguments{
\item{banff_dataset}{A tibble object.}
}
\value{
xxx xxx
}
\description{
This function takes in input a dataset to evaluate the quality of it.
}
\examples{
{

library(fabR)
banff_file <- system.file("extdata", "example.xlsx", package = "banffIT")
banff_dataset <- read_excel_allsheets(banff_file)
banff_dataset_evaluate(banff_dataset)

}

}
