% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_algorithm.R
\name{banff_dataset_evaluate}
\alias{banff_dataset_evaluate}
\title{Assess a Banff dataset}
\usage{
banff_dataset_evaluate(banff_dataset)
}
\arguments{
\item{banff_dataset}{A tibble object.}
}
\value{
A list of tibble objects giving information on the assessment of the dataset.
}
\description{
This function takes a dataset and evaluates its format and content based
on the accepted format specified in the data dictionary.
}
\examples{
{

library(fabR)
input_dataset <- system.file("extdata", "example.xlsx", package = "banffIT")
banff_dataset <- read_excel_allsheets(input_dataset)[1,]
banff_dataset_evaluate(banff_dataset)

}

}
