% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_algorithm.R
\name{calculate_adequacy}
\alias{calculate_adequacy}
\title{Calculate adequacy from glomeruli and arteries}
\usage{
calculate_adequacy(banff_dataset)
}
\arguments{
\item{banff_dataset}{A tibble object.}
}
\value{
A tibble object which has two columns, adequacy_calculated and adequacy_input.
}
\description{
If glomeruli and arteries are provided, it calculates adequacy from them.
}
\examples{
{

library(fabR)
banff_file <- system.file("extdata", "example.xlsx", package = "banffIT")
banff_dataset <- read_excel_allsheets(banff_file)
calculate_adequacy(banff_dataset)

}

}
