% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_algorithm.R
\name{calculate_adequacy}
\alias{calculate_adequacy}
\title{Calculate adequacy of each biopsy from glomeruli and arteries}
\usage{
calculate_adequacy(banff_dataset)
}
\arguments{
\item{banff_dataset}{A tibble object.}
}
\value{
A tibble object with two columns: the calculated adequacy (adequacy_calculated)
and the adequacy specified in input (adequacy_input).
}
\description{
This function calculates adequacy of each biopsy (i.e., each observation) based on glomeruli and
arteries variables (if provided).
}
\examples{
{

library(fabR)
input_dataset <- system.file("extdata", "example.xlsx", package = "banffIT")
banff_dataset <- read_excel_allsheets(input_dataset)
calculate_adequacy(banff_dataset)

}

}
