% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{get_banff_dictionary}
\alias{get_banff_dictionary}
\title{Get the banff data dictionary}
\usage{
get_banff_dictionary(
  which = NULL,
  language = "label:en",
  detail = FALSE,
  .test_TA = FALSE
)
}
\arguments{
\item{which}{Indicates which banff dictionary to get. Either "input" or
"output". If NULL, both banff dictionary are provided.}

\item{language}{Optional input allowing the user to get the diagnosis in a
specific language. Options are "label:en" (default), "label:fr", "label:de",
"label:sp", "label:nl", "label:jp", "label:in".}

\item{detail}{Whether the output should include temporary variables generated
in the process or not. FALSE by default.}

\item{.test_TA}{Temporary parameter to test with 'ta' variable.}
}
\value{
A list of tibbles representing meta data used in the process. The metadata
are the list of variables used, and their associated categories, if any.
}
\description{
This function gets the data dictionary used to control the consistancy of
the banff dataset.
}
\examples{
{

  get_banff_dictionary()

}

}
