% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{get_banff_dictionary}
\alias{get_banff_dictionary}
\title{gets the banff dictionary which is used in the process}
\usage{
get_banff_dictionary(
  which = NULL,
  language = "label:en",
  detail = FALSE,
  .test_TA = FALSE
)
}
\arguments{
\item{which}{Indicates which banff dictionary to get. Either "input" or
"output". If NULL, the whole banff dictionary is provided.}

\item{language}{optional input allowing user to get their diagnosis in a
specific language. Options are "label:en" (default), "label:fr", "label:de",
"label:sp", "label:nl", "label:jp", "label:in".}

\item{detail}{Whether the output should include temporary variables generated
in the process or not. FALSE by default.}

\item{.test_TA}{Temoporary parameter to test with 'ta' variable.}
}
\value{
A list of tibble representing meta data used in the process. The meta
data is the list of variables used in the process, and their associated
categories if any.
}
\description{
gets the banff dictionary which is used in the process
}
\examples{
{

  get_banff_dictionary()

}

}
