% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{get_banff_example}
\alias{get_banff_example}
\title{Get a dataset example}
\usage{
get_banff_example(version = NULL)
}
\arguments{
\item{version}{A character string referring the version of Banff classification.
The most recent classification is the default.
Options are "2022" (default), "2017".}
}
\value{
A tibble representing the dataset used in the the vignette as an example.
}
\description{
This function gets the dataset used in the vignette as an example.
}
\examples{
{

  get_banff_example()

}

}
